fix(canvas): address sketchpad api review
- isolate sketchpad views, serializers and domain modules
- use typed compile jobs and record timestamps
- separate operation handlers and shared-field merging
- require annotated reads and separate hydrated replay serializers
- derive read and write schemas from shared builders
- keep actor response fields compatible with desktop clients

Generated-By: PostHog Desktop
Task-Id: 617b374c-f922-473b-96a7-f61b557c16ea
