+++ pyproject.toml
@@ -586,6 +586,13 @@ ignore_imports = [
     "products.canvas.backend.presentation.serializers -> products.canvas.backend.artifacts",
     "products.canvas.backend.presentation.serializers -> products.canvas.backend.contract",
     "products.canvas.backend.presentation.serializers -> products.canvas.backend.models",
+    "products.canvas.backend.presentation.sketchpad.serializers -> products.canvas.backend.models",
+    "products.canvas.backend.presentation.sketchpad.serializers -> products.canvas.backend.sketchpad.records",
+    "products.canvas.backend.presentation.sketchpad.serializers -> products.canvas.backend.sketchpad.schema",
+    "products.canvas.backend.presentation.sketchpad.views -> products.canvas.backend.models",
+    "products.canvas.backend.presentation.sketchpad.views -> products.canvas.backend.sketchpad.log",
+    "products.canvas.backend.presentation.sketchpad.views -> products.canvas.backend.sketchpad.compiler",
+    "products.canvas.backend.presentation.sketchpad.views -> products.canvas.backend.sketchpad.records",
     "products.canvas.backend.presentation.views -> products.canvas.backend.actions",
     "products.canvas.backend.presentation.views -> products.canvas.backend.build_service",
     "products.canvas.backend.presentation.views -> products.canvas.backend.capabilities",
